1. SafeDelete, SafeDeleteArray, IsSafeWindow 매크로 추가 및 적용 2. int 를 모두 long으로 변경 3. Debug 버튼에 있던 기능을 '잔고' 버튼으로 이동시킴. (정식적용) 4. Layout 조정. Grid의 폭을 늘리고 오른쪽 컨트롤들 재배치.

+++ ABotItem.h
@@ -22,7 +22,7 @@ public:
 	void Init();
 
 public:
-	int				m_index;			//순번.
+	long			m_index;			//순번.
 	eItemState		m_eitemState;		//프로세스 상태.
 	CString			m_strCode;			//코드
 	CString			m_strName;			//이름
